requirements.txt
PyYAML~=6.0.2
requests~=2.32.3
aiohttp~=3.11.17